api_usersnack/requirements.txt
psycopg2
flask
Flask-SQLAlchemy
Flask-Migrate
Flask-Script
Flask-WTF
sqlalchemy
pyyaml